Style link, and inline styles for performance
Previously the link inherited styles. On my site this meant little contrast.

I've styled the link in the blue, and specified an underline with a border on hover/focus.

As this is a banner that's above-the-fold, the styles should be inline to prevent the CSS being a render-blocking resource. So I've done this too.